+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Change thunk assembly magic numbers/sentinel values to sequences that are guaranteed to not be emitted by the compiler.
   Thanks to @Dasaav-dsv for the help.
-- Move most of the arch/feature compile_error checks into the build script for better errors.
+- Move the arch/feature compile_error checks into the build script for better errors.
 - Dual license under Apache-2.0 and MIT.
 
 ## [v2.4.0] - 2025-06-08